Implemented restriction of resource creation and deletion only for the users own profile

+++ controllers/experiences.js
@@ -40,13 +40,12 @@ exports.addExperience = async (req, res) => {
 
   try {
     // Check that the user is associated with the profile the addition is being made to
-    const usersProfile = await Profile.findOne( {email: req.user.email} );
-    if (usersProfile && usersProfile.id === profile) {
+    if (req.user.profileId === profile) {
       const experience = new Experience(experienceObject);
       await experience.save();
       return res.status(201).json(experience);
     } else {
-      return res.status(401).json({ msg: 'Unauthorized'});
+      return res.status(401).json({ msg: 'Unauthorized' });
     }
   } catch (err) {
     console.error(err.message);
@@ -72,9 +71,14 @@ exports.deleteExperiences = async (req, res) => {
 // @access Public
 exports.deleteExperience = async (req, res) => {
   try {
-    // TODO: Check that the user is associated with the profile the addition is being made to
-    await Experience.findByIdAndDelete(req.params.id);
-    res.status(200).json({ msg: `Experience ${req.params.id} deleted` });
+    // Check that the user is associated with the profile the change is being made to
+    const experience = await Experience.findById(req.params.id);
+    if (req.user.profileId === education.profile.toString()) {
+      await Experience.findByIdAndDelete(req.params.id);
+      res.status(200).json({ msg: `Experience ${req.params.id} deleted` });
+    } else {
+      res.status(401).json({ msg: 'Unauthorized' });
+    }
   } catch (err) {
     console.error(err.message);
     res.status(500).send('Server Error');
